+++ models/hybrid_search_retreiver.py
@@ -20,7 +20,6 @@ See: https://python.langchain.com/docs/modules/model_io/llms/llm_caching
 import glob
 import os
 import textwrap
-from typing import List
 
 # pinecone integration
 import pinecone
@@ -44,24 +43,22 @@ from langchain.vectorstores.pinecone import Pinecone
 from pinecone_text.sparse import BM25Encoder
 
 # this project
-from models.const import Credentials
+from models.const import Config, Credentials
 
 
 ###############################################################################
 # initializations
 ###############################################################################
-DEFAULT_MODEL_NAME = "text-davinci-003"
+DEFAULT_MODEL_NAME = Config.OPENAI_PROMPT_MODEL_NAME
 pinecone.init(api_key=Credentials.PINECONE_API_KEY, environment=Credentials.PINECONE_ENVIRONMENT)
 set_llm_cache(InMemoryCache())
 
 
 class TextSplitter:
     """
-    Custom text splitter that add metadata to the Document object
+    Custom text splitter that adds metadata to the Document object
     which is required by PineconeHybridSearchRetriever.
     """
-
-    # ...
 
     def create_documents(self, texts):
         """Create documents"""
@@ -74,16 +71,16 @@ class TextSplitter:
 
 
 class HybridSearchRetriever:
-    """Sales Support Model (SSM)."""
+    """Hybrid Search Retriever (OpenAI + Pinecone)"""
 
     # prompting wrapper
     chat = ChatOpenAI(
         api_key=Credentials.OPENAI_API_KEY,
         organization=Credentials.OPENAI_API_ORGANIZATION,
-        cache=True,
-        max_retries=3,
-        model="gpt-3.5-turbo",
-        temperature=0.0,
+        cache=Config.OPENAI_CHAT_CACHE,
+        max_retries=Config.OPENAI_CHAT_MAX_RETRIES,
+        model=Config.OPENAI_CHAT_MODEL_NAME,
+        temperature=Config.OPENAI_CHAT_TEMPERATURE,
     )
 
     # embeddings
@@ -111,22 +108,6 @@ class HybridSearchRetriever:
         llm = OpenAI(model=model)
         retval = llm(prompt.format(concept=concept))
         return retval
-
-    def fit_tf_idf_values(self, corpus: List[str]):
-        """Fit TF-IDF values.
-        1. Fit the BM25 encoder on the corpus
-        2. Encode the corpus
-        3. Store the encoded corpus in Pinecone
-        """
-        corpus = ["foo", "bar", "world", "hello"]
-
-        # fit tf-idf values on your corpus
-        self.bm25_encoder.fit(corpus)
-
-        # persist the values to a json file
-        self.bm25_encoder.dump("bm25_values.json")
-        self.bm25_encoder = BM25Encoder().load("bm25_values.json")
-        self.bm25_encoder.fit(corpus)
 
     def load(self, filepath: str):
         """
@@ -201,9 +182,9 @@ class HybridSearchRetriever:
         document_texts = [doc.page_content for doc in documents]
         leader = textwrap.dedent(
             """\
-            You can assume that the following is true,
-            and you should attempt to incorporate these facts
-            in your response:
+            \n\nYou can assume that the following is true.
+            You should attempt to incorporate these facts
+            into your response:\n\n
         """
         )